# docs/swagger.yaml
definitions:
  authentication.LoginRequest:
    properties:
      password:
        description: |-
          รหัสผ่านที่ใช้ในการเข้าสู่ระบบ
          required: true
        example: password123
        type: string
        x-order: "1"
      username:
        description: |-
          ชื่อผู้ใช้ที่ใช้ในการเข้าสู่ระบบ
          required: true
        example: johndoe
        type: string
        x-order: "0"
    required:
    - password
    - username
    type: object
  authentication.LoginResponse:
    properties:
      expires_after:
        description: ระยะเวลาที่โทเค็นจะหมดอายุหลังจากการสร้าง
        example: 5 minutes
        type: string
        x-order: "4"
      expires_at:
        description: เวลาที่โทเค็นหมดอายุ
        example: "2024-07-19T09:43:39Z"
        type: string
        x-order: "3"
      token:
        description: โทเค็นที่ใช้ในการยืนยันตัวตน
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJqb2huX2RvZSIsImV4cCI6MTY2NzQ2OTg4NywiaWF0IjoxNjY3NDY5NzI3fQ.SvJv9_XERK8TSEhK9y7BX34vxIZYz3ndY8bX_MSRZGQ
        type: string
        x-order: "2"
    type: object
  authentication.RefreshTokenRequest:
    properties:
      refresh_token:
        description: โทเค็นที่ต้องการ Refresh เวลาการใช้งาน
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJqb2huX2RvZSIsImV4cCI6MTY2NzQ2OTg4NywiaWF0IjoxNjY3NDY5NzI3fQ.SvJv9_XERK8TSEhK9y7BX34vxIZYz3ndY8bX_MSRZGQ
        type: string
        x-order: "0"
    type: object
  authentication.RefreshTokenResponse:
    properties:
      expires_at:
        description: เวลาที่โทเค็นหมดอายุ
        example: "2024-07-19T09:43:39Z"
        type: string
        x-order: "2"
      token:
        description: โทเค็นที่ใช้ในการยืนยันตัวตน
        type: string
        x-order: "1"
    type: object
  authentication.RegisterRequest:
    properties:
      customer_id:
        example: 1
        type: integer
        x-order: "0"
      description:
        example: Customer description
        type: string
        x-order: "4"
      name:
        example: John Doe
        type: string
        x-order: "1"
      password:
        example: password123
        type: string
        x-order: "3"
      username:
        example: johndoe
        type: string
        x-order: "2"
    required:
    - customer_id
    - name
    - password
    - username
    type: object
...
  responses.BadRequestResponseSwagger:
    properties:
      message:
        description: ข้อความสถานะของ response
        example: Bad Request
        type: string
        x-order: "3"
      status:
        description: รหัสสถานะของ response
        example: 400
        type: integer
        x-order: "1"
      success:
        description: สถานะของ response
        example: false
        type: boolean
        x-order: "0"
      timestamp:
        description: เวลาที่ response ถูกสร้างขึ้น
        example: "2025-01-01T00:00:00.000000000Z"
        type: string
        x-order: "2"
    type: object
...
  responses.InternalServerErrorResponseSwagger:
    properties:
      message:
        description: ข้อความสถานะของ response
        example: Internal Server Error
        type: string
        x-order: "3"
      status:
        description: รหัสสถานะของ response
        example: 500
        type: integer
        x-order: "1"
      success:
        description: สถานะของ response
        example: false
        type: boolean
        x-order: "0"
      timestamp:
        description: เวลาที่ response ถูกสร้างขึ้น
        example: "2025-01-01T00:00:00.000000000Z"
        type: string
        x-order: "2"
    type: object
...
  responses.SuccessResponseSwagger-authentication_LoginResponse:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/authentication.LoginResponse'
        description: ข้อมูลที่ต้องการส่งกลับ
        x-order: "4"
      message:
        description: ข้อความสถานะของ response
        example: OK
        type: string
        x-order: "3"
      status:
        description: รหัสสถานะของ response
        example: 200
        type: integer
        x-order: "1"
      success:
        description: สถานะของ response
        example: true
        type: boolean
        x-order: "0"
      timestamp:
        description: เวลาที่ response ถูกสร้างขึ้น
        example: "2025-01-01T00:00:00.000000000Z"
        type: string
        x-order: "2"
    type: object
  responses.SuccessResponseSwagger-authentication_RefreshTokenResponse:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/authentication.RefreshTokenResponse'
        description: ข้อมูลที่ต้องการส่งกลับ
        x-order: "4"
      message:
        description: ข้อความสถานะของ response
        example: OK
        type: string
        x-order: "3"
      status:
        description: รหัสสถานะของ response
        example: 200
        type: integer
        x-order: "1"
      success:
        description: สถานะของ response
        example: true
        type: boolean
        x-order: "0"
      timestamp:
        description: เวลาที่ response ถูกสร้างขึ้น
        example: "2025-01-01T00:00:00.000000000Z"
        type: string
        x-order: "2"
    type: object
...
  responses.UnauthorizedResponseSwagger:
    properties:
      message:
        description: ข้อความสถานะของ response
        example: Unauthorized
        type: string
        x-order: "3"
      status:
        description: รหัสสถานะของ response
        example: 401
        type: integer
        x-order: "1"
      success:
        description: สถานะของ response
        example: false
        type: boolean
        x-order: "0"
      timestamp:
        description: เวลาที่ response ถูกสร้างขึ้น
        example: "2025-01-01T00:00:00.000000000Z"
        type: string
        x-order: "2"
    type: object
...
info:
  contact: {}
  description: For eZVeiw Lite Back Office Project.
  title: eZView Lite Back Office API
  version: "1.0"
paths:
  /v1/auth/register/user/customer:
    post:
      consumes:
      - application/json
      description: |-
        API นี้ใช้สำหรับการลงทะเบียนผู้ใช้ใหม่สำหรับลูกค้าและคืนค่า JWT token ซึ่ง Response จะมีข้อมูลดังนี้:

        - **token**: JWT token ที่ได้รับจากการลงทะเบียน
        - **expires_at**: เวลาที่ token หมดอายุ
        - **expires_after**: ระยะเวลาที่ token หมดอายุหลังจากนี้ (60 นาที)

        Customer ID ได้จาก API /v1/customers/individual หรือ /v1/customers/company

        **หมายเหตุ:** กรุณาใช้ข้อมูลของ Customer ID, User Name และ Password ที่ถูกต้องเพื่อลงทะเบียนผู้ใช้ใหม่
      parameters:
      - description: Register Request
        in: body
        name: registerRequest
        required: true
        schema:
          $ref: '#/definitions/authentication.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: การลงทะเบียนสำเร็จ
          schema:
            $ref: '#/definitions/responses.SuccessResponseSwagger-authentication_LoginResponse'
        "400":
          description: การร้องขอไม่ถูกต้อง
          schema:
            $ref: '#/definitions/responses.BadRequestResponseSwagger'
        "500":
          description: ข้อผิดพลาดภายในเซิร์ฟเวอร์
          schema:
            $ref: '#/definitions/responses.InternalServerErrorResponseSwagger'
      summary: ลงทะเบียนผู้ใช้ใหม่สำหรับลูกค้า
      tags:
      - Authentication
...
  /v1/login:
    post:
      consumes:
      - application/json
      description: |-
        API นี้ใช้สำหรับการเข้าสู่ระบบของผู้ใช้และคืนค่า JWT token ซึ่ง Response จะมีข้อมูลดังนี้:

        - **token**: JWT token ที่ได้รับจากการเข้าสู่ระบบ
        - **expires_at**: เวลาที่ token หมดอายุ
        - **expires_after**: ระยะเวลาที่ token หมดอายุหลังจากนี้ (60 นาที)

        **หมายเหตุ:** กรุณาใช้ข้อมูลของ User Name และ Password ที่ถูกต้องเพื่อรับ Token ในการใช้งาน API อื่น ๆ
      parameters:
      - description: Login Request
        in: body
        name: loginRequest
        required: true
        schema:
          $ref: '#/definitions/authentication.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: การเข้าสู่ระบบสำเร็จ
          schema:
            $ref: '#/definitions/responses.SuccessResponseSwagger-authentication_LoginResponse'
        "400":
          description: การร้องขอไม่ถูกต้อง
          schema:
            $ref: '#/definitions/responses.BadRequestResponseSwagger'
        "401":
          description: ข้อมูลเข้าสู่ระบบไม่ถูกต้อง
          schema:
            $ref: '#/definitions/responses.UnauthorizedResponseSwagger'
        "500":
          description: ข้อผิดพลาดภายในเซิร์ฟเวอร์
          schema:
            $ref: '#/definitions/responses.InternalServerErrorResponseSwagger'
      summary: เข้าสู่ระบบ
      tags:
      - Authentication
...
  /v1/refresh-token:
    post:
      consumes:
      - application/json
      description: |
        API นี้ใช้สำหรับรีเฟรช JWT token โดยใช้ Refresh token ที่ได้รับจากการเข้าสู่ระบบเพื่อสร้าง Access token ใหม่ ซึ่ง Response จะมีข้อมูลดังนี้:

        - **token**: JWT token ที่ได้รับจากการรีเฟรช
        - **expires_at**: เวลาที่ token หมดอายุ
      parameters:
      - description: Refresh Token Request
        in: body
        name: refreshToken
        required: true
        schema:
          $ref: '#/definitions/authentication.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: การรีเฟรช Token สำเร็จ
          schema:
            $ref: '#/definitions/responses.SuccessResponseSwagger-authentication_RefreshTokenResponse'
        "400":
          description: การร้องขอไม่ถูกต้อง
          schema:
            $ref: '#/definitions/responses.BadRequestResponseSwagger'
        "401":
          description: ข้อมูลเข้าสู่ระบบไม่ถูกต้อง
          schema:
            $ref: '#/definitions/responses.UnauthorizedResponseSwagger'
        "500":
          description: ข้อผิดพลาดภายในเซิร์ฟเวอร์
          schema:
            $ref: '#/definitions/responses.InternalServerErrorResponseSwagger'
      summary: รีเฟรช JWT token
      tags:
      - Authentication
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"